backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
pillow
python-multipart
reportlab
pyjwt
torch
transformers
huggingface_hub
